update boj c 11382

+++ 이종윤/C/11382.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main(void ) {
-    int a,b,c;
-    scanf("%d %d %d",&a,&b,&c);
-    printf("%d",a+b+c);
+    long long a,b,c;
+    scanf("%lld %lld %lld",&a,&b,&c);
+    printf("%lld",a+b+c);
 }
